Updated backend file structure for more organized routing.
